move shared stuff to another file
